Add files via upload
This dataset was acquired from the website https://www.fc.up.pt/addi/project.html, and is the result of a collaboration amongst doctors from the Pedro Hispano Hospital and researchers from the University of Porto, University of Aveiro and Instituto Superior Técnico. As a surgeon and academic researcher in the area of melanoma and soft tissue tumors, melanoma and its related data is of great interest. I have led clinical trials investigating the efficacy of treatments for melanoma, and believe access to data is essential in order to understand interpret, and treat disorders of the skin.

The original source for the database can be accessed here: https://www.fc.up.pt/addi/ph2%20database.html
The above link provides a description of data and associated labelling. Data is also hosted at https://www.kaggle.com/jamesgoydos/melanoma-skin-lesion-id-ph2-data for notebook and prediction developments.